GFI now works correctly
