+++ adagucserverEC/CMakeLists.txt
@@ -141,6 +141,7 @@ add_library(
     CDPPGoes16Metadata.cpp
     CCreateTiles.h
     CCreateTiles.cpp
+    CUniqueRequests/CUniqueRequests.cpp
     testadagucserver.cpp
 )
