+++ adagucserverEC/CUniqueRequests/CURConstants.h
@@ -1,0 +1,2 @@
+#define CCUniqueRequests_MAX_DIMS 255
+typedef std::map<int, CT::string *>::iterator it_type_dimvalindex;
